Cast float num steps to int.

+++ experiment.py
@@ -60,10 +60,10 @@ def Dataset(params, kind=None):
     step = 0.5
     if kind == "large":
         applied_steps = np.concatenate(
-            [np.linspace(-4, -2, 2 / step, endpoint=False), np.linspace(2 + step, 4, 2 / step)]
+            [np.linspace(-4, -2, int(2 / step), endpoint=False), np.linspace(2 + step, 4, int(2 / step))]
         )
     else:
-        applied_steps = np.linspace(-2, 2, (2 * 2) / step + 1)
+        applied_steps = np.linspace(-2, 2, int((2 * 2) / step) + 1)
 
     params.update(load_abb=True, shuffle=True)
     collect_dataset(
